[ananas-20] Resolves clobbering of forward/reverse strand k-mers.
Resolves #20. Thanks to @ananth-pallaseni for the big hint that allowed me to
crack this... When we were running the `mergeCanon` function to merge
canonical and anti-canonical k-mers, the `mergeCanon` function was flipping
the strands on anti-canonical k-mers. This caused k-mers to be coerced to
canonical if we had a k-mer that was seen both in the forward and reverse
strand. To resolve this, we just remove the `mergeCanon` function, and use
the `merge` function. This also eliminates a `reduceByKey`, which should
improve performance.